requirements.txt
fastapi
uvicorn
python-docx
PyMuPDF
spacy
sentence-transformers
streamlit
transformers
